2nd release of finance changes by Atique Ur Rasheed and Osama.

+++ akf_accounts/utils/encumbrance/enc_material_request.py
@@ -3,44 +3,57 @@ from frappe.utils import get_link_to_form, fmt_money
 from akf_accounts.akf_accounts.doctype.donation.donation import get_currency_args
 from erpnext.accounts.utils import get_company_default
 
-def validate_donor_balance(self):
-	if(self.is_new()): return
-	if(not get_company_default(self.company, "custom_enable_accounting_dimensions_dialog", ignore_validation=True)): 
-		self.set("program_details", [])
+def validate_donor_balance(doc, method=None):
+	self = doc
+	if(self.is_new()): 
 		return
 
-	donor_balance = sum([d.actual_balance for d in self.program_details])
-	item_amount = sum([d.amount for d in self.items])
+	if(self.encumbrance):
 
-	if(not self.program_details):
-		frappe.throw("Balance is required to proceed further.", title='Donor Balance')
-	if(item_amount> donor_balance):
-		frappe.throw(f"Item amount: <b>Rs.{fmt_money(item_amount)}</b> exceeding the available balance: <b>Rs.{fmt_money(donor_balance)}</b>.", title='Donor Balance')
+		if(not get_company_default(self.company, "custom_enable_accounting_dimensions_dialog", ignore_validation=True)): 
+			self.set("custom_program_details", [])
+			return
 
-def make_encumbrance_material_request_gl_entries(self):
-	if(not get_company_default(self.company, "custom_enable_accounting_dimensions_dialog", ignore_validation=True)): return
-	validate_donor_balance(self)
-	args = frappe._dict({
-			'doctype': 'GL Entry',
-			'posting_date': self.transaction_date,
-			'transaction_date': self.transaction_date,
-			'against': f"Material Request: {self.name}",
-			'against_voucher_type': 'Material Request',
-			'against_voucher': self.name,
-			'voucher_type': 'Material Request',
-			'voucher_no': self.name,
-			'voucher_subtype': 'Receive',
-			'remarks': f"Encumbrance of material request, {self.material_request_type}",
-			'is_opening': 'No',
-			'is_advance': 'No',
-			'company': self.company,
-		})
-	amount = sum([d.amount for d in self.items])
-	for row in self.program_details:
-		difference_amount = amount if(row.actual_balance>=amount) else (amount - row.actual_balance)
-		amount = amount - difference_amount
-		make_debit_gl_entry(args, row, difference_amount)
-		make_credit_gl_entry(self.company, args, row, difference_amount)
+		donor_balance = sum([d.actual_balance for d in self.custom_program_details])
+		item_amount = sum([d.amount for d in self.items])
+
+		if(not self.custom_program_details):
+			frappe.throw("Balance is required to proceed further.", title='Donor Balance')
+		
+		if(item_amount> donor_balance):
+			frappe.throw(f"Item amount: <b>Rs.{fmt_money(item_amount)}</b> exceeding the available balance: <b>Rs.{fmt_money(donor_balance)}</b>.", title='Donor Balance')
+
+def make_encumbrance_material_request_gl_entries(doc, method=None):
+	self = doc
+	if(self.encumbrance):
+		
+		if(get_company_default(self.company, "custom_enable_accounting_dimensions_dialog", ignore_validation=True)): 
+			
+			validate_donor_balance(self)
+				
+			args = frappe._dict({
+				'doctype': 'GL Entry',
+				'posting_date': self.transaction_date,
+				'transaction_date': self.transaction_date,
+				'against': f"Material Request: {self.name}",
+				'against_voucher_type': 'Material Request',
+				'against_voucher': self.name,
+				'voucher_type': 'Material Request',
+				'voucher_no': self.name,
+				'voucher_subtype': 'Receive',
+				'remarks': f"Encumbrance of material request, {self.material_request_type}",
+				'is_opening': 'No',
+				'is_advance': 'No',
+				'company': self.company,
+			})
+			
+			amount = sum([d.amount for d in self.items])
+			
+			for row in self.custom_program_details:
+				difference_amount = amount if(row.actual_balance>=amount) else (amount - row.actual_balance)
+				amount = amount - difference_amount
+				make_debit_gl_entry(args, row, difference_amount)
+				make_credit_gl_entry(self.company, args, row, difference_amount)
 
 def make_debit_gl_entry(args, row, amount):
 	cargs = get_currency_args()
@@ -54,6 +67,7 @@ def make_debit_gl_entry(args, row, amount):
 		'subservice_area': row.pd_subservice_area,
 		'product': row.pd_product,
 		'project': row.pd_project,
+		'fund_class': row.pd_fund_class,
 		'donor': row.pd_donor,
 		'debit': amount,
 		'debit_in_account_currency': amount,
@@ -76,6 +90,7 @@ def make_credit_gl_entry(company, args, row, amount):
 		'subservice_area': row.pd_subservice_area,
 		'product': row.pd_product,
 		'project': row.pd_project,
+		'fund_class': row.pd_fund_class,
 		'donor': row.pd_donor,
 		'credit': amount,
 		'credit_in_account_currency': amount,
@@ -86,6 +101,9 @@ def make_credit_gl_entry(company, args, row, amount):
 	doc.insert(ignore_permissions=True)
 	doc.submit()
 	
-def cancel_encumbrance_material_request_gl_entries(self):
-	if(frappe.db.exists('GL Entry', {'against_voucher': self.name})):
-		frappe.db.sql(f""" Delete from `tabGL Entry` where against_voucher = '{self.name}' """)
+def cancel_encumbrance_material_request_gl_entries(doc, method=None):
+	self = doc
+	if(self.encumbrance):
+		if(frappe.db.exists('GL Entry', {'against_voucher': self.name})):
+			frappe.db.sql(f""" Delete from `tabGL Entry` where against_voucher = '{self.name}' """)
+
